fix: Issue #2 - Change case of Element directory reference in Mobilizer Event Index and View files
The Production server, as a Linux box is case sensitive and to that end
requires references to files to also be case sensitive. The CakePHP
format for directories within the Template/Element subdirectory is to
have an initial capital letter. To correspond to this format, the
Element reference was changed in the view templates.